feat: optimize toc anchor experience
