+++ packages/core/src/theme-default/logic/index.ts
@@ -1,6 +1,7 @@
 export { usePrevNextPage } from './usePrevNextPage';
 export { useEditLink } from './useEditLink';
 export { useSidebarData } from './useSidebarData';
+export { useHiddenNav } from './useHiddenNav';
 export { useLocaleSiteData } from './useLocaleSiteData';
 export { setup, bindingAsideScroll, scrollToTarget } from './sideEffects';
 export { usePathUtils } from './usePathUtils';
